src/backend/requirements.txt
Flask==2.0.2
Flask-PyMongo==2.3.0
bleach==4.1.0
ptvsd==4.3.2 # Required for debugging.
Werkzeug==2.2.2
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-pymongo 
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests

